refactor(types): типы данных формы вынесены в свой регистратор
`FormTypesProvider` тянул на себе всё сразу — от обхода конфигурации до
преобразования реквизитов в данные формы (`java:S1448`, `java:S1200`). Первый
разрез: `FormDataTypesRegistrar` забирает то, во что превращается реквизит и
табличная часть объекта, вместе с картами уже заведённых типов. Провайдер
обращается к нему за типами реквизитов и строкой коллекции, а
`ConfigurationTypesProvider` регистрирует зеркала табличных частей напрямую.

Двуязычные имена из метаданных (`neutral`, `bilingual`) переехали в словарь
`FormPlatformTypes` — ими пользуются оба.
